refactor: move helper functions to data_tools
