+++ src/interaction_manager.py
@@ -145,37 +145,3 @@ class InteractionManager:
                 return os.path.join(root, folder)
 
         return None
-
-    # TODO: Move methods to a proper helper module
-    @staticmethod
-    def capitalize_dict(
-        obj: dict[str, str],
-        *,
-        key: bool = True,
-        value: bool = True,
-        delimiter: str = " ",
-        join_sep: str = " ",
-    ) -> dict[str, str]:
-        new_dict: dict[str, Any] = {}
-
-        for k, v in obj.items():
-            key_words: list[str] = []
-            value_words: list[str] = []
-            key_str: str = k
-            value_str: str = v
-
-            if key:
-                for word in k.split(delimiter):
-                    key_words.append(word.capitalize())
-
-                key_str = join_sep.join(key_words)
-
-            if value:
-                for word in v.split(delimiter):
-                    value_words.append(word.capitalize())
-
-                value_str = join_sep.join(value_str)
-
-            new_dict[key_str] = value_str
-
-        return new_dict
